+++ app/dashboard/clients/lib/types/client.type.ts
@@ -1,0 +1,30 @@
+import { ClientStatus } from './client-status.type';
+import { LivingArrangements } from './living-arrangements.type';
+import { TypeOfGuardianship } from './type-of-guardianship.type';
+import { UserGender } from './user-gender.type';
+import { WealthStatus } from './wealth-status.type';
+
+export type Client = {
+  _id: string;
+  guardianId: string;
+  gender: UserGender;
+  title?: string;
+  firstname: string;
+  lastname: string;
+  birthday: string;
+  avatar?: string;
+  localCourt?: string;
+  caseNumber?: string;
+  scopeOfDuties?: string[];
+  guardianshipStartedAt: string;
+  guardianshipEndedAt?: string;
+  livingArrangement?: LivingArrangements;
+  wealthStatus?: WealthStatus;
+  typeOfGuardianship?: TypeOfGuardianship;
+  isGuardianshipTakenOver: boolean;
+  prevGuardianType?: string;
+  prevGuardianshipStartedAt?: string;
+  createdAt?: string;
+  updatedAt?: string;
+  status: ClientStatus;
+};