+++ app/dashboard/clients/lib/types/clients-table-col-keys.ts
@@ -1,0 +1,7 @@
+export type ClientTableColKey =
+  | 'fullname'
+  | 'guardianshipStartedAt'
+  | 'guardianshipEndedAt'
+  | 'livingArrangement'
+  | 'wealthStatus'
+  | 'status';