+++ app/dashboard/clients/lib/types/living-arrangements.type.ts
@@ -1,0 +1,1 @@
+export type LivingArrangements = 'inpatient' | 'outpatientEquivalent' | 'otherLivingArrangement';