+++ app/dashboard/clients/lib/types/type-of-guardianship.type.ts
@@ -1,0 +1,11 @@
+export type TypeOfGuardianship =
+  | 'professionalGuardianship'
+  | 'voluntaryGuardianship'
+  | 'supplementaryGuardianship'
+  | 'estateCustodianship'
+  | 'contactCustodianship'
+  | 'proceedingsGuardianship'
+  | 'proceduralAssistant'
+  | 'executorOfAWill'
+  | 'guardianship'
+  | 'healthcareProxy';