+++ app/lib/types/api-response.types.ts
@@ -1,0 +1,4 @@
+export type APIResponse<D = any, M = any> = {
+  data: D;
+  meta: M;
+};