+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - Feb 28, 2024.
+
+### Added
+
+- Add clients table.
+- Connect clients table with API.
+- Add create clients form.
+- Connect create clients form with API and clients page UI.
+- Add a bunch of new ui components.
+- New user profile badge (now with dropdown).
+
+### Changed
+
+- Adapt to new API response format and add pagination utils. New api res format is {data, meta} object.
+
 ## [0.1.0] - Feb 21, 2024.
 
 ### Added